selectors/simple-selectors/attribute-selectors/css3-selectors-lang-050.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Selectors: [lang="en-GB"], lang="en-US"</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/selectors/#attribute-selectors'>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link href="../../../resources/testharness.css" rel="stylesheet" type="text/css">
<meta name="assert" content="A lang= value and a lang attribute value will NOT match if their region subtags differ.">
<style type='text/css'>
.test div { width: 50px; }
#box[lang='en-GB'] { width: 100px; }
</style>
</head>
<body>



<div class="test"><div id="box" lang="en-US">&#xA0;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, " ");
</script>

<div id='log'></div>

</body>
</html>
